airflow/requirements.txt
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4


pyyaml==6.0.1
colorlog

apache-airflow==2.7.3
psycopg2-binary
asyncpg

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib


loguru==0.7.2